requirements.txt
Django>=4.2.0
Pillow>=10.0.0
qrcode>=7.4.2
python-dotenv>=1.0.0
django-tailwind>=3.6.0
django-browser-reload>=1.12.0
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
whitenoise>=6.0.0
gunicorn>=21.0.0
django-compressor>=4.0
django-libsass>=0.9
django-debug-toolbar>=4.0.0
# Production dependencies
psycopg2-binary==2.9.7
dj-database-url==2.1.0
python-decouple==3.8